+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased](https://github.com/orisai/nette-console/compare/1.3.1...HEAD)
 
+### Added
+
+- `ConsoleExtension`
+	- `name` - allows dynamic parameter
+	- `version` - allows dynamic parameter
+
 ## [1.3.1](https://github.com/orisai/nette-console/compare/1.3.0...1.3.1) - 2023-08-19
 
 ### Fixed